mmc: omap_hsmmc: limit max_segs with the EDMA DMAC
The EDMA DMAC has a hardware limitation that prevents supporting
scatter gather lists with any number of segments. Since the EDMA
DMA Engine driver sets the maximum segments to 16, we do the
same.

TODO: this will be replaced once the DMA Engine API supports an
API to query the DMAC's segment size limit.

Signed-off-by: Matt Porter <mporter@ti.com>
Signed-off-by: Pascal Huerst <pascal.huerst@gmail.com>