usb: musb: dsps: reduce musb instance to one
Currently multiple phy's of the same type are not supported, hence
reduce musb instances to one. Even without this, it was observed that
both instances were working (by luck), but this holds good iff wrapper
is part of Image. And it is not correct for both controller's to be
associated with same phy, here it was working because phy is a nop one.
And having wrapper as a module and rmmod'ing would crash.

Signed-off-by: Afzal Mohammed <afzal@ti.com>