net: phy: at803x: Add support for hardware reset
The AT803x will enter a FIFO error mode if a packet is transmitted while
the cable is unplugged. This severe hardware issue is acknowledged by
the vendor, and the only proposed solution is to conduct a hardware
reset via the external pin. There is apparantly no way to fix up the
state from the register set.

This patch adds support for reading a gpio reset property from the DT
node of the phy. If found, this gpio descriptor is used to apply a
hardware reset each time a 'link down' condition is detected.
All relevant registers are read out before, and written back after the
reset cycle.

Doing this every time the link might seem like overkill, but there is
unfortunately no way of figuring out whether the PHY is in such a
lock-up state. Hence, this is the only way of reliably fixing up things.

Signed-off-by: Daniel Mack <zonque@gmail.com>