ARM: OMAP2+: Hookup amx3xx PM code into OMAP builds
With all the requisite changes in place we can now enable basic
PM support for am33xx and am437x. This patch updates the various
OMAP files to enable suspend-resume to mem and standby on am335x
and am437x.

Because the suspend resume functionality is different on these
platforms than other OMAP platforms due to the need for M3
firmware and an IPC channel to be in place, separate PM ops
are used instead of omap_pm_ops.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>