ASoC: cs4271: Change chache_type from rbtree to flat
We have seen issues after suspend, while syncing back the
register cache to the device. There are bugs in the rbtree
caching functions, which cause wrong values to be written to
the registers. Since rbtree is not really a benefit for devices
using 8 bit adresses, we just change the cache type to flat,
which is O(1) in this case anyways.