arm: dts: imx7d.dtsi: add (for now commented out) 1.2GHz operating point
We add this operating point but leave it commended out for now. Because
the current batch of chips runs fine on 1.2GHz but we got told by NXP
that the chips are not spec'd for this frequency. Later when we will
have a batch which is specified for 1.2GHz, we probably need to
implement some kind of detection mechanism or have a different
devicetree where this operating point is enabled.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>