clk: clk-divider: add CLK_DIVIDER_LAZY write flag
Some dividers (like the one on imx7) will create a glitch if you write to
the divider register, even if it is the same value. For this reason we
add the CLK_DIVIDER_LAZY_WRITE flag which will first read back the current
value and only write a changed value to the register.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>