mach-imx: clk-imx7d: utilize new CLK_DIVIDER_LAZY_WRITE flag for audio clocks
Without this flag we were getting glitches on the MCLK and BCLK line when
changing the PLL. This did not seem to be an issue with the I2S codec, but
it created dropouts on the SPDIF receiver. It turns out the source of the
issue is when you write to any of the divider registers, it will create
this glitch, even if you just write the same value. With this flag the new
value will only be written if it is different then the current one.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>