dma: imx-sdma: fixup residue calculation
buf_tail has the range from 0 to including (num_bd - 1), so
(num_bd - buf_tail) can never be lower than 1 therefore the minimum for
the residues is period_len. This is not correct and breaks GoogleCast and
the snd_pcm_delay() calculations. We fix this by subtraction 1 from num_bd,
so if we are on the last buffer the residue is only bd->mode.count. I'm
pretty sure that this is how it is supposed to be.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>