PCI: imx6: return 1 in pci_imx_phy_pll_locked() if lock was successful
Before this the function was only used to wait for the PLL lock and if
the lock was not successfull the rest of the code continied as if the
PLL was locked, this even led to a kernel hang later on if the PLL was
not acutally locked.

In addition to that the new initialization sequence needs to know if the
lock succeeded or not so a retry can be made if not.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>