pcm-dummy: add support for sue,nmute
Up to now dummy codec had a support for mute line = muted when high. On
streamkitprime we have nmute line = muted when low. Add a new flag
sue,nmute which changes behavior of mute-gpio to nmute

This is currently a workaround, later when we cleanup the audio driver and
the pcm-dummy codec we will use the new gpio framework which has the
GPIO_ACTIVE_LOW and GPIO_ACTIVE_HIGH flags. See
Documentation/gpio/board.txt

Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>