ASoC: davinci-mcasp: Only disable inactive serializer
As a side effect of the following commit, the active TX
serializer may get disabled and it may result in distorted
audio output.

ASoC: davinci-mcasp: Add support for multichannel playback
2952b27e2e463b28d5c0f04000f96b968137ca42

For example, if a 4 channel I2S playback with two TX serializers
is activated. Later on, if a recording of 2 channels, with only 1 RX
serializer is started, which will also disable one of the TX
serializer because max_active_serializers is only calculated for
RX (recording) stream. This patch fixes the issue.

Signed-off-by: Vishal Thanki <vishal.thanki@teufel.de>